+++ src/main/java/org/nodens2k/tin/validation/GreeceTinValidator.java
@@ -1,0 +1,22 @@
+package org.nodens2k.tin.validation;
+
+import java.util.regex.Pattern;
+import org.jetbrains.annotations.Contract;
+
+public final class GreeceTinValidator extends AbstractCountryTinValidator {
+
+  public static final CountryTinValidator INSTANCE = new GreeceTinValidator();
+
+  private static final Pattern TIN_REGEX = Pattern.compile("^[0-9]{9}$");
+
+  private GreeceTinValidator() {
+    super("GR", "GRC", "300");
+  }
+
+  @Contract(value = "null -> false", pure = true)
+  @Override
+  public boolean isValid(String tin) {
+    tin = sanitise(tin);
+    return tin != null && TIN_REGEX.matcher(tin).matches();
+  }
+}